Pushing For Connor to Pull...this is PWM config CRASHING head

+++ Software/CCS/RoverMotherboard/roveIncludes/roveWareHeaders/roveHardwareAbstraction.h
@@ -48,7 +48,7 @@ void pinMode(int pin, int mode);
  */
 void digitalWrite(int pin, int val);
 
-PWM_Handle rovePWMInit(PWM_Handle pwm_index, int period_in_microseconds);
+PWM_Handle rovePWMInit(PWM_Handle pwm_index, uint16_t period_in_microseconds);
 
 void analogWrite(PWM_Handle pin, int val);
